fix: Relocate and map admin components
This commit includes the relocation of all components related to the `admin` part of the builder and other related improvements such:
- Map the files using the `@admin`, `@styles`, and `@images` alias.
- Clean-up the alias list to keep it simple and easier to understand.
