+++ src/assets/scss/builder/_builder-area.scss
@@ -34,7 +34,7 @@
 			height: 56px;
 			display: block;
 			margin-bottom: $spacing-lg;
-			background-image: url('../../assets/img/placeholder-image.svg');
+			background-image: url('@images/placeholder-image.svg');
 			background-repeat: no-repeat;
 			background-size: contain;
 			background-position: center;
