+++ src/components/Admin/Builder/BuilderSidebar.vue
@@ -34,5 +34,5 @@ defineEmits(['add-element', 'download']);
 </script>
 
 <style lang="scss" scoped>
-@forward '../../assets/scss/builder/builder-sidebar';
+@forward '@styles/builder/builder-sidebar';
 </style>
