+++ src/components/__test__/UIRadioImage.test.js
@@ -1,7 +1,7 @@
 import { expect, describe, it } from 'vitest';
 import { mount } from '@vue/test-utils';
 
-import UIRadioImage from '../UI/UIRadioImage.vue';
+import UIRadioImage from '@admin/UI/UIRadioImage.vue';
 
 describe('UIRadioImage', () => {
 	const props = {
